Separate tests into files per test type
